+++ json_serializable/test/kitchen_sink/simple_object.dart
