+++ json_serializable/test/yaml/angular_comp.yaml
