+++ json_serializable/test/yaml/angular_config.yaml
